Add a grid

+++ src/widget.rs
@@ -1,6 +1,6 @@
 use piston_window::{G2d, Context};
 
 pub trait Widget {
-    fn draw(&self, context: Context, gl: &mut G2d, width: u32, height: u32);
+    fn draw(&self, context: Context, gl: &mut G2d, width: f64, height: f64);
 }
